+++ packages/base/eslint.config.mjs
@@ -1,0 +1,3 @@
+import config from "@plsr/eslint-config-nuxt";
+
+export default config;