+++ packages/base/nuxt.config.ts
@@ -1,11 +1,6 @@
 // https://nuxt.com/docs/api/configuration/nuxt-config
 export default defineNuxtConfig({
-  modules: [
-    "@nuxtjs/eslint-module",
-    "@nuxtjs/tailwindcss",
-    "@nuxt/image",
-    "nuxt-icon",
-  ],
+  modules: ["@nuxt/eslint", "@nuxtjs/tailwindcss", "@nuxt/image", "nuxt-icon"],
   devtools: { enabled: true },
   postcss: {
     plugins: {