+++ packages/base/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @plsr/nuxt-layer-base
 
+## 0.0.11
+
+### Patch Changes
+
+- Fix bun workspace
+- Updated dependencies
+  - @plsr/eslint-config-nuxt@0.0.7
+
+## 0.0.10
+
+### Patch Changes
+
+- Fix dependencies issues
+- Updated dependencies
+  - @plsr/eslint-config-nuxt@0.0.6
+
 ## 0.0.9
 
 ### Patch Changes