+++ packages/eslint/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @plsr/eslint-config-nuxt
 
+## 0.0.7
+
+### Patch Changes
+
+- Fix bun workspace
+
+## 0.0.6
+
+### Patch Changes
+
+- Fix dependencies issues
+
 ## 0.0.5
 
 ### Patch Changes